Removed string.h & size_t per requirements.

+++ get_next_line.h
@@ -15,14 +15,13 @@
 
 # include	<unistd.h>
 # include	<stdlib.h>
-# include	<string.h>
 
 void	ft_strdel(void **as);
 char	*ft_strchr(const char *s, int c);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
+char	*ft_substr(char const *s, unsigned int start, unsigned int len);
 char	*ft_strdup(const char *s1);
 size_t	ft_strlen(const char *s);
-void	*ft_memcpy(void *dst, const void *src, size_t n);
+void	*ft_memcpy(void *dst, const void *src, unsigned int n);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*get_next_line(int fd);
